Fix typos, improve and separate classes, types
